fix: refactored assignments location in the project
